Criado a busca dos valores a receber e sua respectiva soma

+++ database/migrations/1578681096942_usuario_schema.js
@@ -13,7 +13,7 @@ class UsuarioSchema extends Schema {
       table.enu("tipo", ["administrador", "consultor"]).notNullable();
       //---- 0 o uruario não está ativo e 1 está ativo ----
       table.boolean("status").notNullable();
-      table.integer("meta").notNullable();
+      table.decimal("meta").notNullable();
       table.integer("porcentagem_comissao").notNullable();
       table.timestamps();
     });
